requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.3
cycler==0.11.0
fastapi==0.79.0
fonttools==4.35.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pydantic==1.9.2
pyparsing==3.0.9
PyPDF2==2.10.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
wordcloud==1.8.2.2
